django/requirements.txt
absl-py==0.6.1
aiofiles==0.4.0
aiohttp==3.4.4
altair==2.2.2
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.7.1
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
autopep8==1.4.2
backcall==0.1.0
bleach==1.5.0
blosc==1.6.2
bokeh==1.0.1
Bottleneck==1.2.1
branca==0.3.1
catboost==0.11.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
constantly==15.1.0
cryptography==2.3.1
cryptography-vectors==2.3.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.29
cytoolz==0.9.0.1
dask==0.20.0
decorator==4.3.0
dill==0.2.8.2
distributed==1.24.0
Django==2.1.2
djangorestframework==3.8.2
entrypoints==0.2.3
eventlet==0.22.1
fastparquet==0.1.6
feather-format==0.4.0
filelock==3.0.10
Flask==1.0.2
Flask-SocketIO==3.0.1
folium==0.6.0
gast==0.2.0
graphviz==0.8.4
greenlet==0.4.15
grpcio==1.12.1
h5py==2.8.0
heapdict==1.0.0
html5lib==0.9999999
hyperlink==18.0.0
idna==2.7
idna-ssl==1.0.0
imageio==2.4.1
imgaug==0.2.6
incremental==17.5.0
ipydatawidgets==4.0.0
ipykernel==5.1.0
ipyscales==0.3.0
ipython==7.1.1
ipython-genutils==0.2.0
ipyvolume==0.5.1
ipywebrtc==0.4.1
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jsonschema==3.0.0a3
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
llvmlite==0.25.0
locket==0.2.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
msgpack==0.5.6
multidict==4.4.2
multiprocess==0.70.6.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.3
notebook==5.7.0
numba==0.40.0
numexpr==2.6.8
numpy==1.15.4
olefile==0.46
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.5.1
parso==0.3.1
partd==0.3.9
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pluggy==0.8.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyarrow==0.11.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
pycurl==7.43.0.2
pydicom==1.1.0
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pypandoc==1.4
pyparsing==2.3.0
Pyro4==4.74
pyrsistent==0.14.5
PySocks==1.6.8
pytest==3.10.0
pytest-flask==0.12.0
pytest-runner==4.2
python-dateutil==2.7.5
python-engineio==2.3.2
python-snappy==0.5.3
python-socketio==1.9.0
pythreejs==2.0.2
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
queuelib==1.5.0
reportlab==3.5.9
requests==2.20.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.1
seaborn==0.9.0
Send2Trash==1.5.0
serpent==1.27
service-identity==17.0.0
SimpleITK==1.1.0
six==1.11.0
sortedcontainers==2.0.5
statsmodels==0.9.0
tables==3.4.4
tabulate==0.8.2
tblib==1.3.2
tensorboard==1.7.0
tensorboardX==1.4
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thrift==0.11.0
toml==0.10.0
toolz==0.9.0
torchvision==0.2.1
tornado==5.1.1
tox==3.5.3
tqdm==4.28.1
traitlets==4.3.2
traittypes==0.2.1
Twisted==18.9.0
typing==3.6.4
urllib3==1.23
uWSGI==2.0.17.1
vincent==0.4.4
virtualenv==16.0.0
vispy==0.5.3
w3lib==1.19.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
yarl==1.2.6
zict==0.1.3
zope.interface==4.6.0